move, extract & comments
