+++ src/main/scala/rover/rdo/ObjectId.scala
@@ -1,0 +1,4 @@
+package rover.rdo
+
+case class ObjectId(asString: String) {
+}
