+++ src/main/scala/rover/rdo/conflict/DiffWithAncestor.scala
@@ -1,35 +1,6 @@
-package rover.rdo.client
+package rover.rdo.conflict
 
 import rover.rdo.state.{AtomicObjectState, RecordedStateModification}
-
-import scala.async.Async.async
-import scala.concurrent.ExecutionContext.Implicits.global
-import scala.concurrent.Future
-
-//FIXME: use hashes instead of Longs/Strings?
-class RdObject[A](var state: AtomicObjectState[A]) {
-
-	// TODO: "is up to date" or "version" methods
-
-	protected final def modifyState(op: AtomicObjectState[A]#Op): Unit = {
-		state = state.applyOp(op)
-//		onStateModified(state)
-	}
-
-	protected def onStateModified(oldState: AtomicObjectState[A]): Future[Unit] = {
-		async {
-
-		}
-	}
-
-	protected final def immutableState: A = {
-		return state.immutableState
-	}
-
-	override def toString: String = {
-		state.toString
-	}
-}
 
 class DiffWithAncestor[A](private val child: AtomicObjectState[A], private val ancestor: AtomicObjectState[A]) {
 
@@ -43,7 +14,7 @@ class DiffWithAncestor[A](private val child: AtomicObjectState[A], private val a
 
 			}
 		}
-		
+
 		throw new RuntimeException("Failed to determine difference with this ancestor")
 	}
