+++ src/main/scala/rover/rdo/conflict/resolve/ConflictResolutionMechanism.scala
@@ -1,4 +1,4 @@
-package rover.rdo.conflict
+package rover.rdo.conflict.resolve
 
 import rover.rdo.state.{AtomicObjectState, BasicAtomicObjectState, MergeOperation}
